requirements.txt
pyspark
pillow
numpy
pandas
matplotlib
transformers
torch
psutil
streamlit